+++ src/Pim/Bundle/CatalogBundle/Entity/ProductAttribute.php
@@ -35,7 +35,7 @@ use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
  *      }
  *  }
  * )
- * 
+ *
  * @ExclusionPolicy("all")
  */
 class ProductAttribute extends AbstractEntityAttribute implements