+++ src/Pim/Bundle/ImportExportBundle/Writer/OrmProductWriter.php
@@ -106,7 +106,7 @@ class OrmProductWriter extends AbstractConfigurableStepElement implements
         $this->addVersionListener->setRealTimeVersioning(false);
         foreach ($items as $item) {
             $this->incrementCount($item);
-            $this->productManager->save($item, true);
+            $this->productManager->save($item, false);
         }
         $this->productManager->handleAllMedia($items);
         $this->stepExecution->setWriteCount(count($items));